add in jest, upgrade babel and start changing the test to jest
