+++ .eslintrc
@@ -29,5 +29,11 @@
     "react": {
       "version": "16.8"
     }
+  },
+  "globals": {
+    "describe": true,
+    "it": true,
+    "jest": true,
+    "expect": true
   }
 }
